+++ src/prefabs/pageWithCardView.js
@@ -1189,7 +1189,7 @@
                         type: 'SIZE',
                         label: 'Width',
                         key: 'width',
-                        value: '',
+                        value: '100%',
                         configuration: {
                           as: 'UNIT',
                         },
@@ -2826,7 +2826,7 @@
                         type: 'SIZE',
                         label: 'Width',
                         key: 'width',
-                        value: '',
+                        value: '100%',
                         configuration: {
                           as: 'UNIT',
                         },
@@ -5798,7 +5798,7 @@
                         type: 'SIZE',
                         label: 'Width',
                         key: 'width',
-                        value: '',
+                        value: '100%',
                         configuration: {
                           as: 'UNIT',
                         },