+++ src/prefabs/pageWithHeaderAndFooter.js
@@ -1197,7 +1197,7 @@
                           type: 'SIZE',
                           label: 'Width',
                           key: 'width',
-                          value: '',
+                          value: '100%',
                           configuration: {
                             as: 'UNIT',
                           },
@@ -3602,7 +3602,7 @@
                           type: 'SIZE',
                           label: 'Width',
                           key: 'width',
-                          value: '',
+                          value: '100%',
                           configuration: {
                             as: 'UNIT',
                           },
@@ -3997,7 +3997,7 @@
                           type: 'SIZE',
                           label: 'Width',
                           key: 'width',
-                          value: '',
+                          value: '100%',
                           configuration: {
                             as: 'UNIT',
                           },